Added delete backup functionality

+++ Main/ReplayParser.ReplaySorter/Backup/IRepository.cs
@@ -9,10 +9,10 @@ namespace ReplayParser.ReplaySorter.Backup
     public interface IRepository<T>
     {
         long Create(T entity);
-        T Get(int id);
+        T Get(long id);
         IEnumerable<T> GetAll();
         IEnumerable<T> Where(Func<T, bool> predicate);
-        void Remove(int id);
+        void Remove(long id);
         void RemoveAll();
     }
 }
